apps/python/pyproject.toml
[project]
name = "py-service"
version = "0.1.0"
description = "Legal document ingestion and AI worker for Jamaican legal data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "typer>=0.12",
  "rich>=13.7",
  "httpx>=0.27",
  "beautifulsoup4>=4.12",
  "pdfplumber>=0.11",
  "SQLAlchemy>=2.0",
  "psycopg2-binary>=2.9",
  "python-dotenv>=1.0",
  "openai>=1.0",
  "pydantic>=2.0",
  "tiktoken>=0.7; platform_system != 'Windows'",  # optional on Windows
]
